rearranged model classes
new setter methods -- getter methods give constant vectors
removed parametrisation so only velocity parameterisation as I/O is implemented
parametrisation in moduli is not common and makes the code more complex.
Changes are allowed only on velocity and density.
If the changes are applied moduli and inverse density get dirty and will be recalculated with prepareForModelling or when the getter functions are called