1st step to prepare the code for the variable Grid
The coordinate object which maps indices to coordinates and vice versa is now created only once in the main file SOFI.cpp.
The next step is to give this class vectors of the coordinates of every 1D index as private members so that a mapping to an irregular grid is possible.

The coordinate object is now part of the interface to create eg. the boundary consitions.
At the final stage every function related to coordinates should get this class.