changed major order from NZ,NY,NX to NY,NZ,NX
this step was necessary to implement a variable grid which consists of multiple regular grids with different grid spacings.
Those grid are connected at the X-Z face so that you can specify different grid spacings with depth (Y)

changing the major order helps to map from 3D coordinates to indeces. every subgrid will be a block in the index vector.