bugfix: solved issues with IndexType unsigned int
One issue is not solved but should not cause problems:
In derivatives.cpp you will find lines like:

"Y-=FDorder/2+1"
if ((Y >= 0) && (Y < NY))

If unsigned int is used for Y, the first condition is always fullfilled.
But if Y gets negative there will be a wrap around so Y would be around (max) int.
In that case the second condition won't be fullfilled because NY has to be much smaller than (max) int.
So this line should work both for int and unsigned int.