first version of 2D/3D elastic varGrid
Extra y direction derivative matrices had to be added for the staggered gridpoints.
At the interfaces a coordinate shift in x or z direction is necessary because staggered gridpoints have the same coordinate than the full gridpoints.

Also a few bugs in the model averaging were fixed. This made an exchange of the reference solution in the 2D acousic varGrid test necessary.