updated node weight calculation for geographer
Node Weights are calculated for each case (2D,3D acoustic,elastic,visco,sh)
Each operation in the forward loop is considered.
A single oparation has a weight which scales with number of gridpoints and is independent from the equationtype
Therfore a summation of the weights of all operation gives the total weight.
2nd order without boundary conditions is the reference weight with the value 1
free surface and interpolation for the vargrid is not included at the moment (should be?)