reduced total memory when using partitioner
By calculating the partion before initialization of the derivative matrices
there is no extra memory needed for the partitioner. The memory of the partotioner will be freed before
initialization. At the moment there is the overhead that Dxb,Dyb and Dzb have to be calculated twice.
This can be avoided.